[feat] Added support for `IndistinctRelationship`
