+++ pyatlan/generator/templates/relations_init.jinja2
@@ -4,6 +4,7 @@ import lazy_loader as lazy
 
 __PYATLAN_ASSET_RELATIONS__ = {
     "relationship_attributes": ["RelationshipAttributes"],
+    "indistinct_relationship": ["IndistinctRelationship"],
     {% for relation in relation_def_infos -%}
     "{{ relation.relationship_def.name | to_snake_case }}": ["{{ relation.relationship_def.name | to_cls_name}}"],
     {% endfor %}
